% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_phewas_description.R
\name{add_phewas_description}
\alias{add_phewas_description}
\title{Adds description of PheWAS code to the table.}
\usage{
add_phewas_description(data, keep.unmatched.rows = F, for.plots = F)
}
\arguments{
\item{data}{Data frame containing the attribute with PheWAS codes. Can additionally be a character vector of PheWAS codes. See details for requirements.}

\item{keep.unmatched.rows}{Keep rows that have no matching PheWAS descriptions? This is passed into keep.y in \code{merge}.}

\item{for.plots}{Renames the output to include "phenotype" and "description". Default is \code{FALSE}.}
}
\description{
Adds a description of the PheWAS code to the results table.
Only differs from the original addPhewasDescription in the name
of the PheWAS code column.
}

