% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table.R
\name{create_phewas_table}
\alias{create_phewas_table}
\title{Calls createPhewasTable.
Pretty much only exists to create consistency in function name formatting
(The other ones in simplePheWAS are formatted with underscores)}
\usage{
create_phewas_table(id.icd9.count, ...)
}
\arguments{
\item{id.icd9.count}{A table containing, for each patient, a list of their 
diagnoses and how often each appears in their records}
}
\value{
A table with rows corresponding to patients and columns corresponding to diagnoses
 Each cell of the table is TRUE/FALSE, depending on whether the patient had the 
 diagnosis appear a sufficient number of times in their record
 Also may contain NA, if exclusions were added (diagnoses too similar to the
 target diagnosis to be useful in an analysis).
}
\description{
Calls createPhewasTable.
Pretty much only exists to create consistency in function name formatting
(The other ones in simplePheWAS are formatted with underscores)
}

