% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_icd9_codes.R
\name{get_icd9_codes}
\alias{get_icd9_codes}
\title{Conversion of ICD-9 codes to PheWAS codes}
\usage{
get_icd9_codes(icd9_query = c("088.81"))
}
\arguments{
\item{icd9_query}{vector of the ICD-9 codes to convert}
}
\value{
None
}
\description{
Prints for the user a list of the PheWAS codes that encompass their
ICD-9 code(s) of interest.
}
\examples{
get_icd9_codes()
get_icd9_codes(c("088.81","335"))

}

